requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-environ==0.11.2
psycopg2-binary==2.9.9
django-allauth==0.61.1
requests==2.32.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0